requirements.txt
django==1.10.4
gunicorn==19.6.0
pillow=4.0.0
